pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "clockify-rag"
version = "5.9.1"
description = "Production-ready RAG system with Apple Silicon support, hybrid retrieval, and modular architecture"
readme = "README.md"
license = "MIT"
authors = ["Clockify Team <support@clockify.me>"]
keywords = [
    "rag",
    "retrieval-augmented-generation",
    "nlp",
    "embeddings",
    "faiss",
    "bm25",
    "apple-silicon",
    "m1",
    "llm",
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Operating System :: MacOS",
    "Operating System :: POSIX :: Linux",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
homepage = "https://github.com/apet97/clockifyRag"
documentation = "https://github.com/apet97/clockifyRag"
repository = "https://github.com/apet97/clockifyRag.git"
packages = [{include = "clockify_rag"}]
include = ["knowledge_helpcenter.md"]

[tool.poetry.dependencies]
python = "^3.11,<3.14"

# HTTP client and networking
requests = "^2.32.0"
urllib3 = "^2.2.0"
httpx = "^0.27.0"

# Numerical computing and core data processing
numpy = "^1.26.0,<2.5.0"
pandas = "^2.1.0,<3.0.0"

# Tokenization / text utilities
tiktoken = "^0.5.0,<0.7.0"
nltk = "^3.8.0,<4.0.0"

# Retrieval (BM25 only - for core functionality)
rank-bm25 = "^0.2.2,<0.3.0"

# LangChain integrations (remote-first Ollama design)
langchain = "^0.2.0,<0.4.0"
langchain-community = "^0.2.0,<0.4.0"
langchain-ollama = "^0.1.0,<0.3.0"

# Document ingestion helpers
PyPDF2 = "^3.0.0,<4.0.0"
beautifulsoup4 = "^4.12.0,<5.0.0"
python-docx = "^0.8.11,<0.9.0"

# Core dependencies for API and CLI
fastapi = "^0.115.0,<0.116.0"
uvicorn = "^0.30.0,<0.31.0"
pydantic = "^2.7.0,<3.0.0"
pydantic-settings = "^2.2.0,<3.0.0"
typer = "^0.12.0,<0.13.0"
# Click 8.2+ changes typer's help formatting; keep <8.2 until typer upgrades
click = "^8.1.0,<8.2.0"
rich = "^13.7.0,<14.0.0"

# Utilities
typing_extensions = "^4.8.0,<5.0.0"
psutil = "^5.9.0,<6.0.0"
asgi-lifespan = {version = ">=1.1.0", optional = true}
pytest = {version = ">=8.2.0", optional = true}
pytest-asyncio = {version = ">=0.24.0", optional = true}
pytest-cov = {version = ">=5.0.0", optional = true}
pytest-xdist = {version = ">=3.5.0", optional = true}
black = {version = ">=24.1.0", optional = true}
ipython = {version = ">=8.20.0", optional = true}
mypy = {version = ">=1.13.0", optional = true}
pre-commit = {version = ">=3.7.0", optional = true}
ruff = {version = ">=0.8.0", optional = true}

# ANN / vector index (FAISS installation handled separately for Apple Silicon)
# On macOS arm64: install via conda (conda install -c conda-forge faiss-cpu)
# On other platforms: Poetry will handle this correctly
# faiss-cpu = {version = "^1.8.0,<1.9.0", markers = "platform_system != 'Darwin' or platform_machine != 'arm64'"}

[tool.poetry.group.embeddings]
optional = true

[tool.poetry.group.embeddings.dependencies]
# Local embeddings (optional - core RAG works with remote Ollama only)
sentence-transformers = "^2.6.0,<3.0.0"
torch = "^2.3.0,<2.6.0"  # Metal/MPS stable window for macOS

[tool.poetry.group.eval]
optional = true

[tool.poetry.group.eval.dependencies]
ragas = ">=0.1.14"
datasets = ">=2.19.0"
evaluate = ">=0.4.0"

[tool.poetry.extras]
dev = [
    "asgi-lifespan",
    "pytest",
    "pytest-asyncio",
    "pytest-cov",
    "pytest-xdist",
    "black",
    "ipython",
    "mypy",
    "pre-commit",
    "ruff",
]

[tool.poetry.scripts]
ragctl = "clockify_rag.cli_modern:app"

# ============================================================================
# Tool Configurations
# ============================================================================

[tool.black]
line-length = 120
target-version = ["py311", "py312"]
include = '\.pyi?$'
extend-exclude = '''
/(
    \.git
  | \.mypy_cache
  | \.pytest_cache
  | \.venv
  | rag_env
  | build
  | dist
  | __pycache__
)/
'''

[tool.ruff]
line-length = 120
target-version = "py311"
exclude = [
    ".git",
    "__pycache__",
    ".venv",
    "rag_env",
    "build",
    "dist",
]

[tool.ruff.lint]
select = [
    "E",     # pycodestyle errors
    "W",     # pycodestyle warnings
    "F",     # pyflakes
]
ignore = [
    "E501",  # line too long (handled by black)
]

[tool.ruff.lint.isort]
known-first-party = ["clockify_rag"]

[tool.mypy]
python_version = "3.11"
check_untyped_defs = true
disallow_any_generics = false
disallow_incomplete_defs = false
disallow_untyped_calls = false
disallow_untyped_decorators = false
disallow_untyped_defs = false
no_implicit_optional = true
warn_redundant_casts = true
warn_return_any = false
warn_unused_configs = true
warn_unused_ignores = true
show_error_codes = true
show_error_context = true
pretty = true
exclude = [
    "^tests/",
    "^rag_env/",
    "^build/",
    "^dist/",
]

[[tool.mypy.overrides]]
module = [
    "requests.*",
    "numpy.*",
    "urllib3.*",
    "sentence_transformers.*",
    "faiss.*",
    "psutil.*",
    "rank_bm25.*",
    "torch.*",
    "hnswlib.*",
    "nltk.*",
    "tiktoken.*",
    "fastapi.*",
    "uvicorn.*",
    "typer.*",
    "pydantic.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short --strict-markers"
asyncio_mode = "auto"
markers = [
    "integration: marks tests as integration tests (deselect with '-m \"not integration\"')",
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]

[tool.coverage.run]
source = ["clockify_rag"]
omit = [
    "tests/*",
    "rag_env/*",
    "*/site-packages/*",
]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]

[tool.isort]
profile = "black"
line_length = 120
known_first_party = ["clockify_rag"]
skip_gitignore = true
